Move hooks to a separate directory
