+++ web-js/.env
@@ -1,4 +1,4 @@
-# REACT_APP_API_TYPE="graphql"
-REACT_APP_API_TYPE="mql"
+REACT_APP_API_TYPE="graphql"
+# REACT_APP_API_TYPE="mql"
 # REACT_APP_API_TYPE="local"
 REACT_APP_REALM_APP_ID="todo-sync-tmikv"
